+++ cmake/target/package/platformio_arduino.cmake
@@ -2,6 +2,7 @@ message("Deploy Target: PlatformIO-Arduino")
 
 set (DEPLOY_FILES ${LIBRARY_SRC}
                   ${LIBRARY_HEADERS}
+                  ${CMAKE_CURRENT_SOURCE_DIR}/src/HsmImpl.hpp
                   ${FILES_SCXML2GEN}
                   ${CMAKE_CURRENT_SOURCE_DIR}/README.md
                   ${CMAKE_CURRENT_SOURCE_DIR}/CHANGELOG.md