+++ examples/00_helloworld/00_helloworld_glibmm.cpp
@@ -5,21 +5,26 @@
 
 using namespace hsmcpp;
 
-enum class States { OFF, ON };
+namespace States {
+    const hsmcpp::StateID_t OFF = 0;
+    const hsmcpp::StateID_t ON = 1;
+}
 
-enum class Events { SWITCH };
+namespace Events {
+    const hsmcpp::EventID_t SWITCH = 0;
+}
 
 int main(const int argc, const char** argv) {
     Glib::init();
     Glib::RefPtr<Glib::MainLoop> mainLoop = Glib::MainLoop::create();
-    HierarchicalStateMachine<States, Events> hsm(States::OFF);
+    HierarchicalStateMachine hsm(States::OFF);
 
-    hsm.registerState(States::OFF, [&hsm](const VariantVector_t& args) {
+    hsm.registerState(States::OFF, [&hsm](const hsmcpp::VariantVector_t& args) {
         printf("Off\n");
         std::this_thread::sleep_for(std::chrono::milliseconds(1000));
         hsm.transition(Events::SWITCH);
     });
-    hsm.registerState(States::ON, [&hsm](const VariantVector_t& args) {
+    hsm.registerState(States::ON, [&hsm](const hsmcpp::VariantVector_t& args) {
         printf("On\n");
         std::this_thread::sleep_for(std::chrono::milliseconds(1000));
         hsm.transition(Events::SWITCH);