+++ examples/00_helloworld/00_helloworld_std.cpp
@@ -3,22 +3,25 @@
 #include <hsmcpp/hsm.hpp>
 #include <thread>
 
-using namespace hsmcpp;
+namespace States {
+    const hsmcpp::StateID_t OFF = 0;
+    const hsmcpp::StateID_t ON = 1;
+}
 
-enum class States { OFF, ON };
-
-enum class Events { SWITCH };
+namespace Events {
+    const hsmcpp::EventID_t SWITCH = 0;
+}
 
 int main(const int argc, const char** argv) {
-    std::shared_ptr<HsmEventDispatcherSTD> dispatcher = std::make_shared<HsmEventDispatcherSTD>();
-    HierarchicalStateMachine<States, Events> hsm(States::OFF);
+    std::shared_ptr<hsmcpp::HsmEventDispatcherSTD> dispatcher = std::make_shared<hsmcpp::HsmEventDispatcherSTD>();
+    hsmcpp::HierarchicalStateMachine hsm(States::OFF);
 
-    hsm.registerState(States::OFF, [&hsm](const VariantVector_t& args) {
+    hsm.registerState(States::OFF, [&hsm](const hsmcpp::VariantVector_t& args) {
         (void)printf("Off\n");
         std::this_thread::sleep_for(std::chrono::milliseconds(1000));
         hsm.transition(Events::SWITCH);
     });
-    hsm.registerState(States::ON, [&hsm](const VariantVector_t& args) {
+    hsm.registerState(States::ON, [&hsm](const hsmcpp::VariantVector_t& args) {
         (void)printf("On\n");
         std::this_thread::sleep_for(std::chrono::milliseconds(1000));
         hsm.transition(Events::SWITCH);