+++ examples/01_trafficlight/01_trafficlight_class.cpp
@@ -3,24 +3,27 @@
 #include "hsmcpp/HsmEventDispatcherGLibmm.hpp"
 #include "hsmcpp/hsm.hpp"
 
-using namespace hsmcpp;
+namespace TrafficLightState {
+    const hsmcpp::StateID_t OFF = 0;
 
-enum class TrafficLightState {
-    OFF,
+    const hsmcpp::StateID_t OPERABLE = 1;
+    const hsmcpp::StateID_t INITIALIZING = 2;
+    const hsmcpp::StateID_t RED = 3;
+    const hsmcpp::StateID_t YELLOW = 4;
+    const hsmcpp::StateID_t GREEN = 5;
+}  // namespace TrafficLightState
 
-    OPERABLE,
-    INITIALIZING,
-    RED,
-    YELLOW,
-    GREEN
-};
+namespace TrafficLightEvent {
+    const hsmcpp::EventID_t POWER_ON = 0;
+    const hsmcpp::EventID_t POWER_OFF = 1;
+    const hsmcpp::EventID_t INIT_DONE = 2;
+    const hsmcpp::EventID_t NEXT_STATE = 3;
+}  // namespace TrafficLightEvent
 
-enum class TrafficLightEvent { POWER_ON, POWER_OFF, INIT_DONE, NEXT_STATE };
-
-class TrafficLight : public HierarchicalStateMachine<TrafficLightState, TrafficLightEvent> {
+class TrafficLight : public hsmcpp::HierarchicalStateMachine {
 public:
     TrafficLight(Glib::RefPtr<Glib::MainLoop> loop)
-        : HierarchicalStateMachine(TrafficLightState::OFF)
+        : hsmcpp::HierarchicalStateMachine(TrafficLightState::OFF)
         , mLoop(loop) {
         registerState(TrafficLightState::OFF, this, &TrafficLight::onOff);
         registerState(TrafficLightState::INITIALIZING, this, &TrafficLight::onInitializing);
@@ -41,21 +44,21 @@ public:
         registerTransition(TrafficLightState::YELLOW, TrafficLightState::GREEN, TrafficLightEvent::NEXT_STATE);
         registerTransition(TrafficLightState::GREEN, TrafficLightState::RED, TrafficLightEvent::NEXT_STATE);
 
-        initialize(std::make_shared<HsmEventDispatcherGLibmm>());
+        initialize(std::make_shared<hsmcpp::HsmEventDispatcherGLibmm>());
     }
 
-    void onOff(const VariantVector_t& args) {
+    void onOff(const hsmcpp::VariantVector_t& args) {
         printf("onOff\n");
         mLoop->quit();
     }
 
-    void onInitializing(const VariantVector_t& args) {
+    void onInitializing(const hsmcpp::VariantVector_t& args) {
         printf("onInitializing\n");
         usleep(1000000);
         transition(TrafficLightEvent::INIT_DONE);
     }
 
-    void onRed(const VariantVector_t& args) {
+    void onRed(const hsmcpp::VariantVector_t& args) {
         static int iteration = 0;
 
         printf("onRed\n");
@@ -64,13 +67,13 @@ public:
         ++iteration;
     }
 
-    void onYellow(const VariantVector_t& args) {
+    void onYellow(const hsmcpp::VariantVector_t& args) {
         printf("onYellow\n");
         usleep(1000000);
         transition(TrafficLightEvent::NEXT_STATE);
     }
 
-    void onGreen(const VariantVector_t& args) {
+    void onGreen(const hsmcpp::VariantVector_t& args) {
         printf("onGreen\n");
         usleep(1000000);
         transition(TrafficLightEvent::NEXT_STATE);