+++ tools/scxml2gen/template.cpp
@@ -3,12 +3,12 @@
 #include "@HPP_FILE@"
 
 @CLASS_NAME@::@CLASS_NAME@()
-    : hsmcpp::HierarchicalStateMachine<@ENUM_STATES@, @ENUM_EVENTS@>(@ENUM_STATES@::@INITIAL_STATE@) {
+    : hsmcpp::HierarchicalStateMachine(@ENUM_STATES@::@INITIAL_STATE@) {
 }
 
 bool @CLASS_NAME@::initialize(const std::shared_ptr<hsmcpp::IHsmEventDispatcher>& dispatcher) {
     configureHsm();
-    return hsmcpp::HierarchicalStateMachine<@ENUM_STATES@, @ENUM_EVENTS@>::initialize(dispatcher);
+    return hsmcpp::HierarchicalStateMachine::initialize(dispatcher);
 }
 
 void @CLASS_NAME@::configureHsm() {
@@ -40,47 +40,47 @@ void @CLASS_NAME@::configureActions() {
     @REGISTER_ACTIONS@
 }
 
-void @CLASS_NAME@::onTransitionFailed(const @ENUM_EVENTS@ event, const hsmcpp::VariantVector_t& args) {
+void @CLASS_NAME@::onTransitionFailed(const hsmcpp::EventID_t event, const hsmcpp::VariantVector_t& args) {
     // do nothing
 }
 
-std::string @CLASS_NAME@::getStateName(const @ENUM_STATES@ state) const {
+std::string @CLASS_NAME@::getStateName(const hsmcpp::StateID_t state) const {
     std::string stateName;
 
 #ifndef HSM_DISABLE_TRACES
     switch(state) {
 ~~~BLOCK:ENUM_STATES_ITEM~~~
         case @ENUM_STATES@::@ENUM_STATES_ITEM@:
-            stateName = "@ENUM_STATES_ITEM@";
+            (void)stateName.assign("@ENUM_STATES_ITEM@");
             break;
 ~~~BLOCK_END~~~
         default:
-            stateName = hsmcpp::HierarchicalStateMachine<@ENUM_STATES@, @ENUM_EVENTS@>::getStateName(state);
+            stateName = hsmcpp::HierarchicalStateMachine::getStateName(state);
             break;
     }
 #else
-    stateName = hsmcpp::HierarchicalStateMachine<@ENUM_STATES@, @ENUM_EVENTS@>::getStateName(state);
+    stateName = hsmcpp::HierarchicalStateMachine::getStateName(state);
 #endif // HSM_DISABLE_TRACES
 
     return stateName;
 }
 
-std::string @CLASS_NAME@::getEventName(const @ENUM_EVENTS@ event) const {
+std::string @CLASS_NAME@::getEventName(const hsmcpp::EventID_t event) const {
     std::string eventName;
 
 #ifndef HSM_DISABLE_TRACES
     switch(event) {
 ~~~BLOCK:ENUM_EVENTS_ITEM~~~
         case @ENUM_EVENTS@::@ENUM_EVENTS_ITEM@:
-            eventName = "@ENUM_EVENTS_ITEM@";
+            (void)eventName.assign("@ENUM_EVENTS_ITEM@");
             break;
 ~~~BLOCK_END~~~
         default:
-            eventName = hsmcpp::HierarchicalStateMachine<@ENUM_STATES@, @ENUM_EVENTS@>::getEventName(event);
+            eventName = hsmcpp::HierarchicalStateMachine::getEventName(event);
             break;
     }
 #else
-    eventName = hsmcpp::HierarchicalStateMachine<@ENUM_STATES@, @ENUM_EVENTS@>::getEventName(event);
+    eventName = hsmcpp::HierarchicalStateMachine::getEventName(event);
 #endif // HSM_DISABLE_TRACES
 
     return eventName;